Run scripting addition commands in System Events.
The prior method of running the scripting additions in the Python process didn’t work (likely correctly) in 10.10, because the app won’t come to the front before it’s promoted to a foreground process by the blocking display_dialog: catch-22.

See http://appscript.sourceforge.net/py-appscript/doc/osax-manual/04_notes.html for an explanation of why I do it this way.

EagleFiler is not used as a target because it is rather difficult to scroll/copy/paste around the document with an app-modal dialog in front.